changed amount from a short to an unsigned int

+++ src/main/java/frc/robot/subsystems/fuelvision/FuelVisionData.java
@@ -6,7 +6,7 @@ public class FuelVisionData implements StructSerializable {
     float distance;
     float angle;
     byte orientation;
-    short amount;
+    int amount;
 
     public static final FuelVisionDataStruct struct = new FuelVisionDataStruct();
 }
